Reduce sync duplication
