+++ ShuffleTask.Application/Services/TaskUpsertedAsyncHandler.cs
@@ -28,7 +28,7 @@ internal class TaskUpsertedAsyncHandler : IAsyncEventHandler<TaskUpsertedEvent>
         try
         {
             TaskItem? existing = await _storage.GetTaskAsync(domainEvent.Task.Id).ConfigureAwait(false);
-            TaskItem incoming = NormalizeIncoming(domainEvent.Task, existing);
+            TaskItem incoming = TaskSyncMerge.NormalizeIncoming(domainEvent.Task, existing);
 
             if (existing == null)
             {
@@ -36,7 +36,7 @@ internal class TaskUpsertedAsyncHandler : IAsyncEventHandler<TaskUpsertedEvent>
                 return;
             }
 
-            if (IsStale(incoming, existing))
+            if (TaskSyncMerge.IsStaleEventTask(incoming, existing))
             {
                 _logger?.LogInformation(
                     "Ignoring stale task update for {TaskId} with version {Version}",
@@ -54,55 +54,4 @@ internal class TaskUpsertedAsyncHandler : IAsyncEventHandler<TaskUpsertedEvent>
         }
     }
 
-    private static bool IsStale(TaskItem incoming, TaskItem existing)
-    {
-        if (incoming.EventVersion > 0 && incoming.EventVersion <= existing.EventVersion)
-        {
-            return true;
-        }
-
-        return incoming.EventVersion <= 0 && incoming.UpdatedAt != default && incoming.UpdatedAt <= existing.UpdatedAt;
-    }
-
-    private static TaskItem NormalizeIncoming(TaskItem task, TaskItem? existing)
-    {
-        TaskItem normalized = task.Clone();
-
-        if (string.IsNullOrWhiteSpace(normalized.Id) && existing != null)
-        {
-            normalized.Id = existing.Id;
-        }
-        else if (string.IsNullOrWhiteSpace(normalized.Id))
-        {
-            normalized.Id = Guid.NewGuid().ToString("n");
-        }
-
-        if (normalized.CreatedAt == default)
-        {
-            normalized.CreatedAt = existing?.CreatedAt ?? DateTime.UtcNow;
-        }
-
-        if (normalized.UpdatedAt == default)
-        {
-            normalized.UpdatedAt = existing?.UpdatedAt ?? DateTime.UtcNow;
-        }
-
-        if (normalized.EventVersion <= 0)
-        {
-            normalized.EventVersion = (existing?.EventVersion ?? 0) + 1;
-        }
-
-        if (!string.IsNullOrWhiteSpace(normalized.UserId))
-        {
-            normalized.DeviceId = null;
-        }
-        else
-        {
-            normalized.UserId = existing?.UserId;
-            normalized.DeviceId = string.IsNullOrWhiteSpace(normalized.DeviceId)
-                ? existing?.DeviceId ?? Environment.MachineName
-                : normalized.DeviceId.Trim();
-        }
-        return normalized;
-    }
 }
